room or inventory apis

+++ InnovaStay/src/main/java/com/training/Repository/RoomRepository.java
@@ -3,5 +3,5 @@ package com.training.Repository;
 import com.training.Entity.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
 }
